Your Name: Matt
Subject Line: Big push for styling

What Happened:

Fixed the columns to display properly
Removes Reset in favor of normalize.
Uses a local version of jQuery for development mode
main.js includes a change to reset the image back to generic on a "no" click
Lots of other styling help including with the <ul>'s to appear on one line when possible.